Refactored the Vote page and added minor improvements
